backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
python-dotenv==1.0.0
email-validator==2.1.0
gunicorn==21.2.0
